==> scripts/7_map.Rmd <==
---
title: "7_CVSH: Map"
output: html_notebook
---

```{r}
#library(remotes)
#remotes::install_github("wmgeolab/rgeoboundaries")
```

```{r}
library(rgdal)
library(sf)
library(tidyverse)
library(ggspatial)
library(raster)
library(ggplot2)
library(rnaturalearth)
library(rnaturalearthdata)
library(elevatr)
library(rgeoboundaries)
library(rnaturalearthhires)
```

shapefiles for major rivers, creeks, lakes, and reservoirs are available from the USGS National Hydrography Dataset (NHD)[https://data.cnra.ca.gov/dataset/nhd-major-features] and here[https://www.sciencebase.gov/catalog/item/61f8b88dd34e622189c32890]

Zoom into: xlim = c(-124.00, -120.00), ylim = c(37.00, 41.00)

### Hatchery Lat/Long from Google Maps
* Major rivers - Sacramento and San Joaquin Rivers
* CH- Battle Creek, Shasta dam/lake:  40.400914, -122.144753
* FRH- Feather River, Oroville dam/lake: 39.518446, -121.553234
* MRH- Mokelumne River, Camanche dam/lake: 38.225070, -121.024417
* NH- American River, Nimbus dam/Natoma lake:  38.633636, -121.224950

```{r}
# hatchery locations
hatcheries <- data.frame(Hatchery = c("CH", "FRH", "MRH", "NH"),
                              lat = 
                           c(40.400914,39.518446,38.225070,38.633636),
                              long = 
                           c(-122.144753,-121.553234,-121.024417,
                                            -121.224950))

# san francisco location
sf <- data.frame(name = "San Francisco", lat = 37.77251260768446, long= -122.43143434432544)
```


National Hydrology Dataset - Major Rivers and Creeks

```{r}
NHD1 <- readOGR(dsn = "../additional/NHD_Major_Rivers_and_Creeks/NHD_Major_Rivers_and_Creeks.shp", stringsAsFactors = F)
```

Extract relevant coordinates from NHD dataset
* str_detect(gnis_name, "Battle Creek")
* str_detect(gnis_name, "Feather")
* str_detect(gnis_name, "Mokelumne")
* str_detect(gnis_name, "Nimbus")
* str_detect(gnis_name, "San Joaquin")
* str_detect(gnis_name, "Sacramento")

```{r}
crks <- as.data.frame(NHD1@data[["gnis_name"]]) %>% 
  distinct()
saveNHD1 <- as.data.frame(NHD1@data[["gnis_name"]]) %>% 
  mutate(gnis_name = NHD1@data[["gnis_name"]]) %>% 
  dplyr::select(gnis_name) %>% 
  distinct() %>% 
  filter(str_detect(gnis_name, "American River|Fork Battle Creek|Feather River|Mokelumne River|Sacramento|San Joaquin")) %>% 
  filter(!str_detect(gnis_name, "Channel|Slough"))
testNHD1 <- NHD1 %>% subset(., gnis_name %in% saveNHD1$gnis_name)
ggNHD1 <-fortify(testNHD1)
```


Help for making elevation maps from here: https://rspatialdata.github.io/elevation.html

```{r}
sf_usa <- ne_states(geounit = "United States of America", returnclass = "sf")
sf_ca <- sf_usa %>% 
  filter(abbrev == "Calif.")

elevation_1 <- elevatr::get_elev_raster(locations = sf_ca, z = 7, clip = "locations")
cropped_elev <- crop(elevation_1, sf_ca)
elevate <- as.data.frame(cropped_elev, xy = TRUE)

colnames(elevate)[3] <- "elevation_value"
elevate <- elevate[complete.cases(elevate), ]

a <-ggplot() +
  geom_raster(data = elevate, aes(x = x, y = y, fill = elevation_value)) +
  geom_sf(data = sf_ca, color = "white", fill = NA) +
  coord_sf(xlim = c(-125.00, -120.3), ylim = c(37.00, 41))+
  scale_fill_gradient(low = "gray90", high = "gray45", aesthetics = "fill")+
  labs(title = "", x = "", y = "", fill = "Elevation (meters)") +
  geom_polygon(data=ggNHD1, aes(x = long, y = lat, group=group), color = "dodgerblue")+
  geom_text(data = hatcheries, aes(long, lat, group = Hatchery, label=Hatchery), size = 4, fontface = "bold", nudge_x = c(-0.45,0.5,0.4,0.3), nudge_y = c(0, -0.1, -0.1, 0))+
  geom_point(data = hatcheries, aes(x = long, y = lat), size=3.25, color="black")+
  geom_point(data = hatcheries, aes(x = long, y = lat, color=Hatchery), size=3)+
  scale_color_manual(values = c("#9379AD", "#e4af62", "#b65d2b", "#79ad9f"))+
  annotate(geom = "text", x = -124.3, y = 38.5, label = "Pacific Ocean", fontface = "italic", color = "grey25", size = 3) +
  annotation_scale(location = "bl", width_hint = 0.35) + 
  annotation_north_arrow(location = "bl", which_north = "true", pad_x = unit(0.25, "in"), pad_y = unit(0.25, "in"), 
                         style = north_arrow_fancy_orienteering) +
  theme(panel.grid.major = element_line(color = gray(.9), linetype = "dashed"), 
        panel.background = element_rect(fill = "aliceblue"), axis.text.x = element_text(angle = 45, hjust = 1, size = 5),
        axis.text.y = element_text(size=5))+
  guides(shape = guide_legend(override.aes = list(size = 1)), color = guide_legend(override.aes = list(size = 1)))+
  theme(legend.title = element_text(size = 3), legend.text = element_text(size = 3), legend.key.size = unit(0.5, 'cm'))
```

```{r}
a
```

```{r}
a1 <- a+
  geom_point(data = sf, aes(x=long, y=lat), size=2)+
  annotate(geom = "text", x = -123.1, y = 37.7, label = "San Francisco", fontface = "italic", color = "grey25", size = 3)
a1
```


This is as good as it's probably going to get, so will clean up the rivers a little in Illustrator. The speckles under MRH will be removed, the Battle Creek forks will be connected to the CH dot, and where the Sacramento/San Joaquin rivers connect to the Pacific will be touched up. 


# Inset Map

```{r}
sf_inset <- fortify(ne_coastline())

rectangle <- tibble(x=c(-124,-120.5,-120.5,-124),y=c(37.5, 37.5, 41, 41),
                    group=c(87.1,87.1,87.1,87.1))

b <- ggplot(data = sf_inset, mapping = aes(x = long, y = lat, group = group)) +
  coord_fixed(1.3) + 
  geom_polygon(color = "gray25", fill = "gray90")+
  coord_sf(xlim = c(-125.00, -65), ylim = c(10, 70)) +
  labs(title = "", x = "", y = "") +
  theme(panel.grid.major = element_line(color = gray(.9), linetype = "dashed"), 
        panel.background = element_rect(fill = "aliceblue"), axis.text.x = element_text(size = 5),
        axis.text.y = element_text(size = 5)) +
  geom_polygon(data=rectangle, aes(x=x,y=y,group=group), color="black", fill ="red")
b
```

It'll be easier to add the inset map to the map figure in Illustrator as well. 

```{r}
ggsave(plot=a1, file="../figures/Fig1a_map.png", width = 20, height = 15, units = "cm", dpi = 800)
ggsave(plot=b, file="../figures/Fig1b_inset.png", width = 20, height = 15, units = "cm", dpi = 800)
```
